+++ test/Telegram.Bot.Tests.Integ/Framework/XunitExtensions/XunitTestCollectionRunnerWithAssemblyFixture.cs
@@ -9,8 +9,8 @@ namespace Telegram.Bot.Tests.Integ.Framework.XunitExtensions
 {
     public class XunitTestCollectionRunnerWithAssemblyFixture : XunitTestCollectionRunner
     {
-        readonly Dictionary<Type, object> _assemblyFixtureMappings;
-        readonly IMessageSink _diagnosticMessageSink;
+        private readonly Dictionary<Type, object> _assemblyFixtureMappings;
+        private readonly IMessageSink _diagnosticMessageSink;
 
         public XunitTestCollectionRunnerWithAssemblyFixture(
             Dictionary<Type, object> assemblyFixtureMappings,
@@ -21,23 +21,34 @@ namespace Telegram.Bot.Tests.Integ.Framework.XunitExtensions
             ITestCaseOrderer testCaseOrderer,
             ExceptionAggregator aggregator,
             CancellationTokenSource cancellationTokenSource)
-            : base(testCollection, testCases, diagnosticMessageSink, messageBus, testCaseOrderer, aggregator, cancellationTokenSource)
+            : base(
+                testCollection,
+                testCases,
+                diagnosticMessageSink,
+                messageBus,
+                testCaseOrderer,
+                aggregator,
+                cancellationTokenSource)
         {
             _assemblyFixtureMappings = assemblyFixtureMappings;
             _diagnosticMessageSink = diagnosticMessageSink;
         }
 
-        protected override async Task<RunSummary> RunTestClassAsync(ITestClass testClass, IReflectionTypeInfo @class, IEnumerable<IXunitTestCase> testCases)
+        protected override async Task<RunSummary> RunTestClassAsync(
+            ITestClass testClass,
+            IReflectionTypeInfo @class,
+            IEnumerable<IXunitTestCase> testCases)
         {
-            // Don't want to use .Concat + .ToDictionary because of the possibility of overriding types,
-            // so instead we'll just let collection fixtures override assembly fixtures.
+            // Don't want to use .Concat + .ToDictionary because of the possibility of overriding
+            // types, so instead we'll just let collection fixtures override assembly fixtures.
             var combinedFixtures = new Dictionary<Type, object>(_assemblyFixtureMappings);
             foreach (var kvp in CollectionFixtureMappings)
             {
                 combinedFixtures[kvp.Key] = kvp.Value;
             }
 
-            // We've done everything we need, so let the built-in types do the rest of the heavy lifting
+            // We've done everything we need, so let the built-in types do the rest of the
+            // heavy lifting
             var runner = new XunitTestClassRunner(
                 testClass,
                 @class,